services/leads/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pyjwt==2.8.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0
